Speeding up loading
When the client goes to get a list of posts, it now receives a list of the post _ids_ from the server, which is a much faster operation. For every post that the user tries to see, the client will check a local ethereal "cache" of id to data and query the server for the post's actual data if it doesn't have it.